Pull LocalStorageTheme higher
