+++ src/client/src/MainRoute.tsx
@@ -5,7 +5,6 @@ import { timer } from 'rxjs'
 import { ConnectionActions } from 'rt-actions'
 import { Environment } from 'rt-components'
 import { AutobahnConnectionProxy } from 'rt-system'
-import { ThemeName, ThemeState } from 'rt-theme'
 
 import { createApplicationServices } from './applicationServices'
 import { getEnvVars } from './config/config'
@@ -59,55 +58,10 @@ export default class MainRoute extends React.Component {
       <React.Fragment>
         <ReduxProvider store={this.store}>
           <Environment.Provider value={this.environment}>
-            <LocalStorageThemeProvider>
-              <Router />
-            </LocalStorageThemeProvider>
+            <Router />
           </Environment.Provider>
         </ReduxProvider>
       </React.Fragment>
-    )
-  }
-}
-
-interface State {
-  themeName: ThemeName
-}
-
-const THEME_STORAGE_KEY = 'themeName'
-
-class LocalStorageThemeProvider extends React.Component<{}, State> {
-  state = {
-    themeName: ThemeName.LIGHT
-  }
-
-  componentDidMount = () => {
-    this.setThemeNameFromStorage()
-    window.addEventListener('storage', this.setThemeNameFromStorage)
-  }
-
-  componentWillUnmount = () => {
-    window.removeEventListener('storage', this.setThemeNameFromStorage)
-  }
-
-  getThemeNameFromStorage = () => localStorage.getItem(THEME_STORAGE_KEY)
-
-  setThemeNameFromStorage = () => {
-    const themeName = this.getThemeNameFromStorage()
-    if (themeName === ThemeName.LIGHT || themeName === ThemeName.DARK) {
-      this.setState({ themeName })
-    }
-  }
-
-  updateLocalStorageThemeName = (name: string) => {
-    localStorage.setItem(THEME_STORAGE_KEY, name)
-    this.setThemeNameFromStorage()
-  }
-
-  render() {
-    return (
-      <ThemeState.Provider name={this.state.themeName} onChange={this.updateLocalStorageThemeName}>
-        {this.props.children}
-      </ThemeState.Provider>
     )
   }
 }
