+++ src/client/src/index.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { BrowserRouter, Route, Switch } from 'react-router-dom'
+import { LocalStorageThemeProvider } from 'rt-theme'
 
 import 'rt-theme'
 
@@ -8,11 +9,13 @@ import MainRoute from './MainRoute'
 import NotificationRoute from './NotificationRoute'
 
 ReactDOM.render(
-  <BrowserRouter>
-    <Switch>
-      <Route path="/notification" component={NotificationRoute} />
-      <Route component={MainRoute} />
-    </Switch>
-  </BrowserRouter>,
+  <LocalStorageThemeProvider>
+    <BrowserRouter>
+      <Switch>
+        <Route path="/notification" component={NotificationRoute} />
+        <Route component={MainRoute} />
+      </Switch>
+    </BrowserRouter>
+  </LocalStorageThemeProvider>,
   document.getElementById('root')
 )
