+++ src/client/src/rt-theme/index.ts
@@ -8,3 +8,4 @@ export { resolvesColor } from './tools'
 export { ThemeProvider } from './ThemeProvider'
 export { ThemeState, ThemeContextValue, ThemeName } from './ThemeState'
 export { default as styled } from './styled'
+export { default as LocalStorageThemeProvider } from './LocalStorageThemeProvider'
